scripts/requirements.txt
spotipy==2.23.0
supabase==2.3.0
requests==2.31.0
resend==0.7.0
python-dotenv==1.0.0
